move lca rule to a new file
